reorganize admin & badge actions
