+++ frontend/src/containers/index.js
@@ -9,5 +9,5 @@ export { default as Dashboard } from './Dashboard.js';
 export { default as Admin } from './Admin.js';
 export { default as AdminRoles } from './AdminRoles.js';
 export { default as AdminUsers } from './AdminUsers.js';
-export { default as BadgesAdmin } from './admin/Badges.js';
+export { default as AdminBadges } from './AdminBadges.js';
 export { default as AdminUserEdit } from './AdminUserEdit.js';
